Auth

+++ src/main/java/com/noahasano/expense/dto/IncomeDTO.java
@@ -1,5 +1,6 @@
 package com.noahasano.expense.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import lombok.Data;
@@ -11,11 +12,12 @@ public class IncomeDTO {
 
     private String title;
 
-    private Integer amount;
+    private BigDecimal amount;
 
     private LocalDate date;
 
     private String category;
 
     private String description;
+    
 }
